Move Population to model package
